+++ test/vanilla/Expected/AcceptanceTests/Constants/constants/models/_models_py3.py
@@ -176,47 +176,35 @@ class ModelAsStringRequiredTwoValueNoDefault(msrest.serialization.Model):
 class NoModelAsStringNoRequiredOneValueDefault(msrest.serialization.Model):
     """NoModelAsStringNoRequiredOneValueDefault.
 
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    :ivar parameter:  Default value: "value1".
-    :vartype parameter: str
+    :param parameter:  The only acceptable values to pass in are None and "value1". The default
+     value is "value1".
+    :type parameter: str
     """
-
-    _validation = {
-        "parameter": {"constant": True},
-    }
 
     _attribute_map = {
         "parameter": {"key": "parameter", "type": "str"},
     }
 
-    parameter = "value1"
-
-    def __init__(self, **kwargs):
+    def __init__(self, *, parameter: Optional[str] = "value1", **kwargs):
         super(NoModelAsStringNoRequiredOneValueDefault, self).__init__(**kwargs)
+        self.parameter = parameter
 
 
 class NoModelAsStringNoRequiredOneValueNoDefault(msrest.serialization.Model):
     """NoModelAsStringNoRequiredOneValueNoDefault.
 
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    :ivar parameter:  Default value: "value1".
-    :vartype parameter: str
+    :param parameter:  The only acceptable values to pass in are None and "value1". The default
+     value is None.
+    :type parameter: str
     """
-
-    _validation = {
-        "parameter": {"constant": True},
-    }
 
     _attribute_map = {
         "parameter": {"key": "parameter", "type": "str"},
     }
 
-    parameter = "value1"
-
-    def __init__(self, **kwargs):
+    def __init__(self, *, parameter: Optional[str] = None, **kwargs):
         super(NoModelAsStringNoRequiredOneValueNoDefault, self).__init__(**kwargs)
+        self.parameter = parameter
 
 
 class NoModelAsStringNoRequiredTwoValueDefault(msrest.serialization.Model):
@@ -262,7 +250,7 @@ class NoModelAsStringRequiredOneValueDefault(msrest.serialization.Model):
 
     All required parameters must be populated in order to send to Azure.
 
-    :ivar parameter: Required.  Default value: "value1".
+    :ivar parameter:  Has constant value: "value1".
     :vartype parameter: str
     """
 
@@ -287,7 +275,7 @@ class NoModelAsStringRequiredOneValueNoDefault(msrest.serialization.Model):
 
     All required parameters must be populated in order to send to Azure.
 
-    :ivar parameter: Required.  Default value: "value1".
+    :ivar parameter:  Has constant value: "value1".
     :vartype parameter: str
     """