+++ test/vanilla/Expected/AcceptanceTests/Validation/validation/models/_models.py
@@ -17,7 +17,7 @@ class ChildProduct(msrest.serialization.Model):
 
     All required parameters must be populated in order to send to Azure.
 
-    :ivar const_property: Required. Constant string. Default value: "constant".
+    :ivar const_property: Constant string. Has constant value: "constant".
     :vartype const_property: str
     :param count: Count.
     :type count: int
@@ -46,9 +46,9 @@ class ConstantProduct(msrest.serialization.Model):
 
     All required parameters must be populated in order to send to Azure.
 
-    :ivar const_property: Required. Constant string. Default value: "constant".
+    :ivar const_property: Constant string. Has constant value: "constant".
     :vartype const_property: str
-    :ivar const_property2: Required. Constant string2. Default value: "constant2".
+    :ivar const_property2: Constant string2. Has constant value: "constant2".
     :vartype const_property2: str
     """
 
@@ -110,12 +110,13 @@ class Product(msrest.serialization.Model):
     :type child: ~validation.models.ChildProduct
     :param const_child: Required. The product documentation.
     :type const_child: ~validation.models.ConstantProduct
-    :ivar const_int: Required. Constant int. Default value: "0".
+    :ivar const_int: Constant int. Has constant value: 0.
     :vartype const_int: int
-    :ivar const_string: Required. Constant string. Default value: "constant".
+    :ivar const_string: Constant string. Has constant value: "constant".
     :vartype const_string: str
-    :ivar const_string_as_enum: Constant string as Enum. Default value: "constant_string_as_enum".
-    :vartype const_string_as_enum: str
+    :param const_string_as_enum: Constant string as Enum. The only acceptable values to pass in are
+     None and "constant_string_as_enum". The default value is None.
+    :type const_string_as_enum: str
     """
 
     _validation = {
@@ -126,7 +127,6 @@ class Product(msrest.serialization.Model):
         "const_child": {"required": True},
         "const_int": {"required": True, "constant": True},
         "const_string": {"required": True, "constant": True},
-        "const_string_as_enum": {"constant": True},
     }
 
     _attribute_map = {
@@ -142,7 +142,6 @@ class Product(msrest.serialization.Model):
 
     const_int = 0
     const_string = "constant"
-    const_string_as_enum = "constant_string_as_enum"
 
     def __init__(self, **kwargs):
         super(Product, self).__init__(**kwargs)
@@ -151,3 +150,4 @@ class Product(msrest.serialization.Model):
         self.image = kwargs.get("image", None)
         self.child = kwargs["child"]
         self.const_child = kwargs["const_child"]
+        self.const_string_as_enum = kwargs.get("const_string_as_enum", None)